fix(ui): Get plugin locale from the plugin API (fixes #22)
Previously, the editor locale was determined using navigator.languages.
This, however, doesn't always match Joplin's locale. This commit
switches to using joplin.settings.globalValue to fetch the app locale.

To apply this locale to the renderer, this commit uses
options.settingValue and a new plugin setting that mirrors the global
Joplin locale option.